+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## 1.18.0 (TBD)
 
+### Snowpark Python API Updates
+
+#### New Features
+
+#### Improvements
+
+### Snowpark pandas API Updates
+
+#### New Features
+
+#### Improvements
+
+- Added partial support for `DataFrame.pivot_table` with no `index` parameter, as well as for `margins` parameter.
+
 ### Snowpark Local Testing Updates
 
 #### Bug Fixes